Reduce memory use of sample array
Time is now saved in samples as 4 byte int and
json buffers are reduced to 500bytes.
This allows sample array to contain up to 2000
samples.

Will be useful when we implement store and send
data at end of test logic.

+++ firmware/libraries/potentiostat/ps_sample.h
@@ -8,7 +8,7 @@ namespace ps
     {
         public:
 
-            uint64_t t;
+            uint32_t t;
             float volt;
             float curr;
             uint8_t chan;
